[modules] geo_mag: move WMM2010 model to math/pprz_geodetic_wmm2010
