+++ sw/airborne/math/pprz_geodetic_wmm2010.h
