+++ sw/airborne/modules/geo_mag/geo_mag.c
@@ -1,13 +1,32 @@
 /*
+ * Copyright (C) 2012  Sergey Krukowski <softsr@yahoo.de>
  *
- * Copyright (C) 2012  Sergey Krukowski <softsr@yahoo.de>.
+ * This file is part of paparazzi.
  *
+ * paparazzi is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2, or (at your option)
+ * any later version.
+ *
+ * paparazzi is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with paparazzi; see the file COPYING.  If not, write to
+ * the Free Software Foundation, 59 Temple Place - Suite 330,
+ * Boston, MA 02111-1307, USA.
+ */
+
+/**
+ * @file modules/geo_mag/geo_mag.c
+ * @brief Calculation of the Geomagnetic field vector based on current GPS fix.
  * This module based on the WMM2010 modell (http://www.ngdc.noaa.gov/geomag/models.shtml).
- *
  */
 
 #include "modules/geo_mag/geo_mag.h"
-#include "modules/geo_mag/wmm2010.h"
+#include "math/pprz_geodetic_wmm2010.h"
 #include "math/pprz_algebra_double.h"
 #include "subsystems/gps.h"
 #include "autopilot.h"
